% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{avoc}
\alias{avoc}
\title{Available Option Chain}
\usage{
avoc(symbol)
}
\arguments{
\item{symbol}{Stock Ticker}
}
\value{
Returns a vector of dates representing the expirations
}
\description{
Finds all available option expirations
}
\examples{
avoc("AAPL")

}
